Fix verify max-lines failures
