+++ src/renderer/src/components/settings/terminal-color-override-group-types.ts
@@ -1,0 +1,6 @@
+import type { TerminalColorOverrides } from '../../../../shared/types'
+
+export type TerminalColorOverrideGroup = {
+  label: string
+  keys: { key: keyof TerminalColorOverrides; label: string; description: string }[]
+}
